Moved components to their folders
