+++ src/components/Climber/Climber.tsx
@@ -1,4 +1,4 @@
-import Button from "../Button"
+import Button from "../Button/Button"
 
 import { sharedStyles } from "../../sharedStyles"
