+++ src/components/Tablet/Tablet.tsx
