+++ src/components/Tablet/TabletDifficulty.tsx
